Update devices list, refactor packages
