+++ .circleci/config.yml
@@ -95,9 +95,9 @@ jobs:
             sudo gcloud firebase test android run \
               --app app/build/outputs/apk/debug/app-debug.apk \
               --test app/build/outputs/apk/androidTest/debug/app-debug-androidTest.apk \
-              --device model=hammerhead,version=23,locale=en,orientation=portrait \
-              --device model=redfin,version=30,locale=en,orientation=portrait \
-              --device model=bluejay,version=32,locale=en,orientation=portrait \
+              --device model=sailfish,version=25,locale=en,orientation=portrait \
+              --device model=oriole,version=31,locale=en,orientation=portrait \
+              --device model=husky,version=34,locale=en,orientation=portrait \
 
   scan-sast-pr:
     parameters:
