+++ app/src/androidTest/java/com/verygoodsecurity/demoapp/utils/matchers/WithCardCVCStateMatcher.kt
@@ -1,20 +1,20 @@
-package com.verygoodsecurity.demoapp.matchers
+package com.verygoodsecurity.demoapp.utils.matchers
 
 import android.view.View
 import androidx.test.espresso.matcher.BoundedMatcher
 import androidx.test.espresso.remote.annotation.RemoteMsgConstructor
-import com.verygoodsecurity.vgscollect.widget.PersonNameEditText
+import com.verygoodsecurity.vgscollect.widget.CardVerificationCodeEditText
 import org.hamcrest.Description
 
-class WithCardHolderStateMatcher @RemoteMsgConstructor internal constructor(
+class WithCardCVCStateMatcher @RemoteMsgConstructor internal constructor(
     var str: String
-) : BoundedMatcher<View?, PersonNameEditText>(PersonNameEditText::class.java) {
+) : BoundedMatcher<View?, CardVerificationCodeEditText>(CardVerificationCodeEditText::class.java) {
 
     override fun describeTo(description: Description) {
         description.appendText("with card holder input field state: ")
     }
 
-    override fun matchesSafely(textView: PersonNameEditText): Boolean {
+    override fun matchesSafely(textView: CardVerificationCodeEditText): Boolean {
         val state = textView.getState()
         return state?.run {
             str.length == contentLength
