+++ app/src/androidTest/java/com/verygoodsecurity/demoapp/utils/matchers/WithVGSEditTextStateMatcher.kt
@@ -1,4 +1,4 @@
-package com.verygoodsecurity.demoapp.matchers.validation
+package com.verygoodsecurity.demoapp.utils.matchers
 
 import android.view.View
 import androidx.test.espresso.matcher.BoundedMatcher
@@ -6,15 +6,18 @@ import androidx.test.espresso.remote.annotation.RemoteMsgConstructor
 import com.verygoodsecurity.vgscollect.widget.VGSEditText
 import org.hamcrest.Description
 
-class WithValidationStateMatcher @RemoteMsgConstructor internal constructor() :
-    BoundedMatcher<View?, VGSEditText>(VGSEditText::class.java) {
+class WithVGSEditTextStateMatcher @RemoteMsgConstructor internal constructor(
+    var str: String
+) : BoundedMatcher<View?, VGSEditText>(VGSEditText::class.java) {
 
     override fun describeTo(description: Description) {
-        description.appendText("with editText input field state: ")
+        description.appendText("VGSEditText input field: ")
     }
 
     override fun matchesSafely(textView: VGSEditText): Boolean {
         val state = textView.getState()
-        return state?.isValid ?: false
+        return state?.run {
+            str.length == contentLength
+        }?:false
     }
 }
