+++ app/src/androidTest/java/com/verygoodsecurity/demoapp/utils/matchers/validation/WithValidationStateMatcher.kt
@@ -1,4 +1,4 @@
-package com.verygoodsecurity.demoapp.matchers
+package com.verygoodsecurity.demoapp.utils.matchers.validation
 
 import android.view.View
 import androidx.test.espresso.matcher.BoundedMatcher
@@ -6,18 +6,15 @@ import androidx.test.espresso.remote.annotation.RemoteMsgConstructor
 import com.verygoodsecurity.vgscollect.widget.VGSEditText
 import org.hamcrest.Description
 
-class WithVGSEditTextStateMatcher @RemoteMsgConstructor internal constructor(
-    var str: String
-) : BoundedMatcher<View?, VGSEditText>(VGSEditText::class.java) {
+class WithValidationStateMatcher @RemoteMsgConstructor internal constructor() :
+    BoundedMatcher<View?, VGSEditText>(VGSEditText::class.java) {
 
     override fun describeTo(description: Description) {
-        description.appendText("VGSEditText input field: ")
+        description.appendText("with editText input field state: ")
     }
 
     override fun matchesSafely(textView: VGSEditText): Boolean {
         val state = textView.getState()
-        return state?.run {
-            str.length == contentLength
-        }?:false
+        return state?.isValid ?: false
     }
 }
